backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
google-genai==1.13.0
python-dotenv==1.1.0
pillow==11.2.1